rename `point_extesions` to `extensions` to match content
