+++ lib/src/layer/tile_layer/tile_range.dart
@@ -3,7 +3,7 @@ import 'dart:math' show Point;
 
 import 'package:flutter_map/src/layer/tile_layer/tile_coordinates.dart';
 import 'package:flutter_map/src/misc/bounds.dart';
-import 'package:flutter_map/src/misc/point_extensions.dart';
+import 'package:flutter_map/src/misc/extensions.dart';
 import 'package:meta/meta.dart';
 
 @immutable
